requirements.txt
faiss-cpu
groq
langchain-groq
langchain
langchain_community
langchain_google_genai
streamlit
python-dotenv
PyPDF2
pypdf
google-cloud-aiplatform>=1.38
huggingface_hub
transformers
torch
accelerate
tiktoken
chromadb
unstructured
numpy
pandas
requests

fastapi
uvicorn[standard]
scikit-learn
